hoang 1

+++ src/main/java/com/example/project_final_2/dto/request/CartItemRequestDTO.java
@@ -16,5 +16,5 @@ public class CartItemRequestDTO {
     private Long productId;
 
     @Positive
-    private long numberItem;
+    private int numberItem;
 }
